# benchdnn: improve performance related to memory operations (fixes MFDNN-13451)