--- a/tests/benchdnn/binary/ref_binary.cpp
+++ b/tests/benchdnn/binary/ref_binary.cpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright 2019-2024 Intel Corporation
+* Copyright 2019-2025 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -45,7 +45,7 @@ void compute_ref(
         const auto idx_B = dst.get_idx(i, broadcast_mask_B);
 
         const bool c_val = prb->is_ternary_op()
-                ? static_cast<bool>(src2.get_elem(idx_A))
+                ? static_cast<bool>(src2.get_f32_elem(idx_A))
                 : false;
 
         float res = compute_binary(