--- a/tests/benchdnn/bnorm/bnorm.cpp
+++ b/tests/benchdnn/bnorm/bnorm.cpp
@@ -632,7 +632,8 @@ int init_ref_memory_args(dnn_mem_map_t &ref_mem_map, dnn_mem_map_t &mem_map,
         // use switch below to define a memory desc for it.
         if (exec_arg != DNNL_ARG_SCRATCHPAD && exec_arg != DNNL_ARG_WORKSPACE) {
             ref_mem_map.emplace(exec_arg,
-                    dnn_mem_t(mem.md_, dnnl_f32, tag::abx, ref_engine));
+                    dnn_mem_t(mem.md_, dnnl_f32, tag::abx, ref_engine,
+                            /* prefill = */ false));
         }
 
         switch (exec_arg) {
@@ -641,7 +642,8 @@ int init_ref_memory_args(dnn_mem_map_t &ref_mem_map, dnn_mem_map_t &mem_map,
                     // Stash for backward which is used in reference code:
                     //     src_hat[i] = (src[i] - mean) / sqrt(var + prb->eps)
                     ref_mem_map.emplace(DNNL_ARG_DST_1,
-                            dnn_mem_t(mem.md_, dnnl_f32, tag::abx, ref_engine));
+                            dnn_mem_t(mem.md_, dnnl_f32, tag::abx, ref_engine,
+                                    /* prefill = */ false));
                 }
                 break;
             case DNNL_ARG_DIFF_SRC: break; // Skip on backward.
@@ -650,13 +652,13 @@ int init_ref_memory_args(dnn_mem_map_t &ref_mem_map, dnn_mem_map_t &mem_map,
             case DNNL_ARG_VARIANCE:
                 if (prb->dir & FLAG_INF) {
                     const dnnl_dims_t dims1d = {prb->ic};
-                    ref_mem_map[exec_arg] = dnn_mem_t(
-                            1, dims1d, dnnl_f32, tag::abx, ref_engine);
+                    ref_mem_map[exec_arg] = dnn_mem_t(1, dims1d, dnnl_f32,
+                            tag::abx, ref_engine, /* prefill = */ false);
                 }
                 break;
             case DNNL_ARG_WORKSPACE: {
-                ref_mem_map[exec_arg]
-                        = dnn_mem_t(mem.md_, dnnl_u8, tag::abx, ref_engine);
+                ref_mem_map[exec_arg] = dnn_mem_t(mem.md_, dnnl_u8, tag::abx,
+                        ref_engine, /* prefill = */ false);
                 break;
             }
             default: break;
@@ -678,8 +680,8 @@ int init_ref_memory_args(dnn_mem_map_t &ref_mem_map, dnn_mem_map_t &mem_map,
     // Reference code uses different kind of workspace. Adjust to ref needs.
     if (ref_mem_map.count(DNNL_ARG_WORKSPACE)) {
         const auto &src_md = ref_mem_map[DNNL_ARG_SRC].md_;
-        ref_mem_map[DNNL_ARG_WORKSPACE]
-                = dnn_mem_t(src_md, dnnl_u8, tag::abx, ref_engine);
+        ref_mem_map[DNNL_ARG_WORKSPACE] = dnn_mem_t(
+                src_md, dnnl_u8, tag::abx, ref_engine, /* prefill = */ false);
     }
 
     return OK;