--- a/tests/benchdnn/bnorm/ref_bnorm.cpp
+++ b/tests/benchdnn/bnorm/ref_bnorm.cpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright 2017-2024 Intel Corporation
+* Copyright 2017-2025 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -47,21 +47,21 @@ void compute_ref_fwd(const prb_t *prb, const args_t &args) {
     const auto &attr = prb->attr;
 
     benchdnn_parallel_nd(C, [&](int64_t c) {
-        float smean = mean.get_elem(c);
-        float svar = var.get_elem(c);
+        float smean = mean.get_f32_elem(c);
+        float svar = var.get_f32_elem(c);
         float sqrt_var = sqrtf(svar + prb->eps);
         float rcp_denom = 1.f / sqrt_var;
-        float gamma = use_sc ? sc.get_elem(c) : 1.f;
-        float beta = use_sh ? sh.get_elem(c) : 0.f;
+        float gamma = use_sc ? sc.get_f32_elem(c) : 1.f;
+        float beta = use_sh ? sh.get_f32_elem(c) : 0.f;
 
         for_(int64_t mb = 0; mb < MB; ++mb)
         for_(int64_t d = 0; d < D; ++d)
         for_(int64_t h = 0; h < H; ++h)
         for (int64_t w = 0; w < W; ++w) {
             auto off = data_off(prb, mb, c, d, h, w);
-            float x_hat = (src.get_elem(off) - smean) * rcp_denom;
+            float x_hat = (src.get_f32_elem(off) - smean) * rcp_denom;
             float res = gamma * x_hat + beta;
-            if (fuse_add_relu) res += src_add.get_elem(off);
+            if (fuse_add_relu) res += src_add.get_f32_elem(off);
             if (fuse_relu && res < 0) res = 0;
             if (need_ws) ws_ptr[off] = !!res;
             maybe_post_ops(attr, res);
@@ -96,8 +96,8 @@ void compute_ref_bwd(const prb_t *prb, const args_t &args) {
     const float MB_SP = MB * D * H * W;
 
     benchdnn_parallel_nd(C, [&](int64_t c) {
-        float rcp_denom = 1.f / sqrtf(var.get_elem(c) + prb->eps);
-        float gamma = use_sc ? sc.get_elem(c) : 1.f;
+        float rcp_denom = 1.f / sqrtf(var.get_f32_elem(c) + prb->eps);
+        float gamma = use_sc ? sc.get_f32_elem(c) : 1.f;
 
         float d_gamma = 0;
         float d_beta = 0;
@@ -107,9 +107,9 @@ void compute_ref_bwd(const prb_t *prb, const args_t &args) {
         for_(int64_t h = 0; h < H; ++h)
         for (int64_t w = 0; w < W; ++w) {
             auto off = data_off(prb, mb, c, d, h, w);
-            float dd = d_dst.get_elem(off);
+            float dd = d_dst.get_f32_elem(off);
             if (fuse_relu && ws.get_elem(off) == 0) dd = 0;
-            d_gamma += dd * src_hat.get_elem(off);
+            d_gamma += dd * src_hat.get_f32_elem(off);
             d_beta += dd;
         }
 
@@ -121,13 +121,13 @@ void compute_ref_bwd(const prb_t *prb, const args_t &args) {
         for_(int64_t h = 0; h < H; ++h)
         for (int64_t w = 0; w < W; ++w) {
             auto off = data_off(prb, mb, c, d, h, w);
-            float dd = d_dst.get_elem(off);
+            float dd = d_dst.get_f32_elem(off);
             if (fuse_relu && ws.get_elem(off) == 0) dd = 0;
             if (fuse_add_relu) d_src_add.set_elem(off, dd);
             float ds = dd;
 
             if (!glob_stats)
-                ds -= (d_beta + src_hat.get_elem(off) * d_gamma) / MB_SP;
+                ds -= (d_beta + src_hat.get_f32_elem(off) * d_gamma) / MB_SP;
 
             d_src.set_elem(off, rcp_denom * ds * gamma);
         }