… tests/benchdnn/conv/conv_dw_fusion.cpp
… tests/benchdnn/conv/ref_conv.cpp
… tests/benchdnn/deconv/deconv.cpp
… tests/benchdnn/deconv/ref_deconv.cpp
… tests/benchdnn/dnnl_common.cpp
… tests/benchdnn/dnnl_common.hpp
… tests/benchdnn/dnnl_memory.cpp
… tests/benchdnn/dnnl_memory.hpp
… tests/benchdnn/eltwise/ref_eltwise.cpp
… tests/benchdnn/gnorm/gnorm.cpp
… tests/benchdnn/gnorm/ref_gnorm.cpp
… tests/benchdnn/graph/custom_driver.cpp
… tests/benchdnn/graph/graph_memory.cpp
… tests/benchdnn/graph/input_displacer.cpp
… tests/benchdnn/ip/ip.cpp
… tests/benchdnn/ip/ref_ip.cpp
… tests/benchdnn/lnorm/lnorm.cpp
… tests/benchdnn/lnorm/ref_lnorm.cpp
… tests/benchdnn/lrn/lrn.cpp
… tests/benchdnn/lrn/ref_lrn.cpp
… tests/benchdnn/matmul/matmul.cpp
… tests/benchdnn/matmul/ref_matmul.cpp
… tests/benchdnn/pool/pool.cpp
… tests/benchdnn/pool/ref_pool.cpp
… tests/benchdnn/prelu/prelu.cpp
… tests/benchdnn/prelu/ref_prelu.cpp
… tests/benchdnn/reduction/reduction.cpp
… tests/benchdnn/reduction/ref_reduction.cpp
… tests/benchdnn/reorder/ref_reorder.cpp
… tests/benchdnn/reorder/reorder.cpp