MFDNN-13451

Thanks to @echeresh, who pivoted all the optimizations from the tracker and provided a PoC for them.

This PR implements 4 out of 6 of specified in the tracker. 2 not included is check_zero_pad on inputs and simplified RNG-related filling changes.

On a list attached to the tracker the speed up looks like this:
Before:
```
total: 76.37s; create_pd: 0.04s (0%); create_prim: 7.55s (10%); fill: 31.24s (41%); execute: 2.17s (3%); compute_ref: 9.06s (12%); compare: 23.82s (31%);
```
After:
```
total: 54.19s; create_pd: 0.04s (0%); create_prim: 7.19s (13%); fill: 19.07s (35%); execute: 2.13s (4%); compute_ref: 8.65s (16%); compare: 15.23s (28%);
```